--- src/pages/Home/Home.tsx
@@ -1,9 +1,0 @@
-const Home = () => {
-  return (
-    <>
-      <h1>Home Page</h1>
-    </>
-  );
-};
-
-export default Home;
